Improved FS functionality by  a fair amount.
